Remove notifications from bugreport generation using combo keys
In Q, we added a new bugreport API that asks for user consent before
showing the "Share" notification for the bugreport. We will be migrating
all bugreporting clients to use the API in R.

In the new API workflow we are moving away from broadcasts and using
callbacks to communicate with dumpstate. Finished bugreport notification
broadcast used to share the final path of the bugreport but API now uses
file descriptors instead. Hence, moving away from broadcasts.

The full bugreport in this case can be retrieved from dumpstate directory
using 'adb pull' or finding and attaching the file directly to gmail or
betterbug.

Remove outputfile flag since it's unused.

Bug: 135186519
Bug: 137825702
Test: Build and flash to the device. Create a bugreport using combo
keys, share notification does not pop up. Full bugreport created in
bugreports/ dir

Change-Id: Ibbfa87449faae559bf08b1ed672bfc523f57f8a4