Re-enable Bluetooth "PhonePolicy"
The "PhonePolicy" is used to determine which Bluetooth proiles become
connected when a device is connected, and what order the connections
happen.

Disabling this will prevent the HID Host CTS test from failing as the
HID Host profile does not become connected automatically after pairing.

In a perfect world, a new "TvPolicy" would be created that is used on
Android TV platforms. But for now, re-enable PhonePolicy...

Bug: 130399746
Test: CTS HID Host test
Change-Id: I66993c75a2a5fa41873b6f327c183f35ec65ac3d