yukawa: boot: don't use vintf_fragments for boot hal
With commit 6a6d32aa374f ("yukawa: rework folder tree"), we have added
the boot control hal as a vintf fragment instead of a separate manifest.

Because we also use 'recovery_available: true", the bootctrl.xml is
installed in 2 places:
- system/etc/vintf/manifest/bootctrl.xml (for recovery)
- vendor/etc/vintf/manifest/bootctrl.xml (for normal Android)

The above breaks checkvintf as following:

> getFrameworkHalManifest: -2147483648 VINTF parse error: Cannot add
> manifest fragment /system/etc/vintf/manifest/bootctrl.xml: Cannot add a
> device manifest to a framework manifest

Fix it by removing the usage of vintf_fragments.
Also, update the filepath for the device fragment to point towards the
correct file.

Test: make TARGET_VIML3=true TARGET_KERNEL_USE=4.19 TARGET_USE_AB_SLOT=true
Fixes: 6a6d32aa374f ("yukawa: rework folder tree")
Change-Id: Ifb3ab5a8c7191f8bbb8a502d289a6521f872ff10
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>